# src/main/java/org/openapitools/api/TransactionsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.0.2).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Account;
import java.util.List;
import org.openapitools.model.Transaction;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2019-07-04T16:05:03.095817700+02:00[Africa/Harare]")

@Validated
@Api(value = "transactions", description = "the transactions API")
public interface TransactionsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    @ApiOperation(value = "Create a new transaction", nickname = "createTransaction", notes = "", authorizations = {
        @Authorization(value = "wallet_auth", scopes = {
            @AuthorizationScope(scope = "write:transaction", description = "modify transaction")
            })
    }, tags={ "Transaction", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK"),
        @ApiResponse(code = 400, message = "Invalid account ID"),
        @ApiResponse(code = 405, message = "Invalid authorisation") })
    @RequestMapping(value = "/transactions",
        consumes = { "application/Json" },
        method = RequestMethod.POST)
    default ResponseEntity<Void> createTransaction(@ApiParam(value = ""  )  @Valid @RequestBody List<Account> body) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    @ApiOperation(value = "Get transaction details", nickname = "getTransactionDetails", notes = "", response = Transaction.class, authorizations = {
        @Authorization(value = "wallet_auth", scopes = {
            @AuthorizationScope(scope = "read:transaction", description = "read transaction")
            })
    }, tags={ "Transaction", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = Transaction.class) })
    @RequestMapping(value = "/transactions/{transactionId}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<Transaction> getTransactionDetails(@ApiParam(value = "",required=true) @PathVariable("transactionId") String transactionId) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    ApiUtil.setExampleResponse(request, "application/json", "{  \"toAccount\" : {    \"accountNumber\" : \"accountNumber\",    \"openingBalance\" : 0  },  \"settled\" : false,  \"fromAccount\" : {    \"accountNumber\" : \"accountNumber\",    \"openingBalance\" : 0  },  \"dateInitiation\" : \"2000-01-23T04:56:07.000+00:00\",  \"dateSettlement\" : \"2000-01-23T04:56:07.000+00:00\",  \"transactionId\" : \"transactionId\"}");
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    @ApiOperation(value = "Get all transactions", nickname = "getTransactions", notes = "", response = Account.class, responseContainer = "List", authorizations = {
        @Authorization(value = "wallet_auth", scopes = {
            @AuthorizationScope(scope = "write:transactions", description = ""),
            @AuthorizationScope(scope = "read:transactions", description = "")
            })
    }, tags={ "Transaction", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Account.class, responseContainer = "List"),
        @ApiResponse(code = 400, message = "Invalid status value") })
    @RequestMapping(value = "/transactions",
        produces = { "application/xml", "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<List<Account>> getTransactions() {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    ApiUtil.setExampleResponse(request, "application/json", "{  \"balance\" : 6,  \"closed\" : false,  \"accountNumber\" : 0,  \"transactions\" : [ \"transactions\", \"transactions\" ],  \"availableBalance\" : 1}");
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/xml"))) {
                    ApiUtil.setExampleResponse(request, "application/xml", "<Account>  <accountNumber>123456789</accountNumber>  <balance>123456789</balance>  <availableBalance>123456789</availableBalance>  <closed>true</closed>  <transactions>aeiou</transactions></Account>");
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    @ApiOperation(value = "Reverse transaction", nickname = "transactionsTransactionIdDelete", notes = "", authorizations = {
        @Authorization(value = "wallet_auth", scopes = {
            @AuthorizationScope(scope = "write:transaction", description = "modify transaction")
            })
    }, tags={ "Transaction", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK"),
        @ApiResponse(code = 402, message = "Invalid authentication") })
    @RequestMapping(value = "/transactions/{transactionId}",
        method = RequestMethod.DELETE)
    default ResponseEntity<Void> transactionsTransactionIdDelete(@ApiParam(value = "",required=true) @PathVariable("transactionId") String transactionId) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}