# swagger.yml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "Swagger Petstore"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "www.virtual-wallet.io"
basePath: "/api"
...
schemes:
- "https"
paths:
...
  /transactions:

    get:
      tags:
      - "Transaction"
      summary: "Get all transactions"
      description: ""
      operationId: "getTransactions"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Account"
        400:
          description: "Invalid status value"
      security:
      - wallet_auth:
        - "write:transactions"
        - "read:transactions"

    post:
      tags:
      - "Transaction"
      summary: "Create a new transaction"
      description: ""
      operationId: "createTransaction"
      consumes:
      - "application/Json"
      parameters:
        - in: "body"
          name: "body"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Account"
      responses:
        200:
          description: OK
        400:
          description: "Invalid account ID"
        405:
          description: "Invalid authorisation"
      security:
        - wallet_auth:
          - "write:transaction"
  /transactions/{transactionId}:

    get:
      tags:
        - "Transaction"
      summary: "Get transaction details"
      operationId: "getTransactionDetails"
      produces:
      - "application/json"
      parameters:
        - name: "transactionId"
          in: "path"
          type: "string"
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Transaction"
      deprecated: false
      security:
        - wallet_auth:
          - "read:transaction"

    delete:
      tags:
        - "Transaction"
      summary: "Reverse transaction"
      parameters:
        - name: "transactionId"
          in: "path"
          type: "string"
          required: true
      responses:
        200:
          description: OK
        402:
          description: "Invalid authentication"
      deprecated: false
      security:
        - wallet_auth:
          - "write:transaction"
  /user/login:
    get:
      tags:
      - "user"
      summary: "Logs user into the system"
      description: ""
      operationId: "loginUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "username"
        in: "query"
        description: "The user name for login"
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "The password for login in clear text"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
          headers:
            X-Rate-Limit:
              type: "integer"
              format: "int32"
              description: "calls per hour allowed by the user"
            X-Expires-After:
              type: "string"
              format: "date-time"
              description: "date in UTC when token expires"
        400:
          description: "Invalid username/password supplied"
...
securityDefinitions:
  wallet_auth:
    type: "oauth2"
    authorizationUrl: "https://www.virtual-wallet.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:user: "modify user"
      read:user: "read user"
      write:account: "modify account"
      read:account: "read your account"
      write:transaction: "modify transaction"
      read:transaction: "read transaction"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Transaction:
    type: "object"
    properties:
      transactionId:
        type: "string"
      fromAccount:
        type: "object"
        properties:
          accountNumber:
            type: "string"
          openingBalance:
            type: "integer"
            format: "int64"
      toAccount:
        type: "object"
        properties:
          accountNumber:
            type: "string"
          openingBalance:
            type: "integer"
            format: "int64"
      dateInitiation:
        type: "string"
        format: "date-time"
      dateSettlement:
        type: "string"
        format: "date-time"
      settled:
        type: "boolean"
        default: false
    xml:
      name: "Transaction"
  Account:
    type: "object"
    properties:
      accountNumber:
        type: "integer"
        format: "int64"
      balance:
        type: "integer"
        format: "int64"
      availableBalance:
        type: "integer"
        format: "int64"
      closed:
        type: "boolean"
        default: false
      transactions:
        type: "array"
        format: "string"
    xml:
      name: "Account"
...